+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ### Changed
 
-- Making identification numbers **strings instead of integers**.
+- BREAKING: Making identification numbers **strings instead of integers**.
   This allows us to represent wildcard `F`s at any position.
 - Parsing of the lower layers (DLL and TPL) have been made less brittle and will
   in more cases register errors in the context instead of raising an exception.
+- BREAKING: Device value in DLL and TPL have been changed from an atom to a struct
+  of type `Exmbus.Parser.Tpl.Device`
 
 ### Added